Fix for sonar
