+++ src/main/java/fr/insee/rmes/bauhaus_services/rdf_utils/RepositoryPublication.java
@@ -1,6 +1,5 @@
 package fr.insee.rmes.bauhaus_services.rdf_utils;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -28,7 +27,6 @@ import org.json.JSONObject;
 import fr.insee.rmes.bauhaus_services.Constants;
 import fr.insee.rmes.config.Config;
 import fr.insee.rmes.exceptions.RmesException;
-import fr.insee.rmes.persistance.ontologies.QB;
 
 /**
  * Getters only get on publication base
@@ -203,36 +201,6 @@ public class RepositoryPublication extends RepositoryUtils{
 		clearStructureAndComponents(structure, REPOSITORY_PUBLICATION_INTERNE);
 	}
 
-	public static void clearStructureAndComponents(Resource structure, Repository repository) throws RmesException {
-		List<Resource> toRemove = new ArrayList<>();
-		try (RepositoryConnection conn = repository.getConnection()){
-			RepositoryResult<Statement> nodes = null;
-			RepositoryResult<Statement> specifications = null;
-			nodes = conn.getStatements(structure, QB.COMPONENT, null, false);
-			while (nodes.hasNext()) {
-				Resource node = (Resource) nodes.next().getObject();
-				toRemove.add(node);
-				specifications = conn.getStatements(node, QB.COMPONENT, null, false);
-				while (specifications.hasNext()) {
-					toRemove.add((Resource) specifications.next().getObject());
-				}
-				specifications.close();
-
-			}
-			nodes.close();
-			toRemove.forEach(res -> {
-				try {
-					RepositoryResult<Statement> statements = conn.getStatements(res, null, null, false);
-					conn.remove(statements);
-					statements.close();
-				} catch (RepositoryException e) {
-					logger.error("RepositoryGestion Error {}", e.getMessage());
-				}
-			});
-		} catch (RepositoryException e) {
-			throw new RmesException(HttpStatus.SC_INTERNAL_SERVER_ERROR, e.getMessage(), "Failure deletion : " + structure);
-		}
-	}
 	private static void clearConceptLinks(Resource concept, RepositoryConnection conn) throws RmesException {
 		List<IRI> typeOfLink = Arrays.asList(SKOS.BROADER, SKOS.NARROWER, SKOS.MEMBER, DCTERMS.REFERENCES,
 				DCTERMS.REPLACES, SKOS.RELATED);
